add read list and new readme
