+++ pruebas/01-reading-list/offneti/.vscode/extensions.json
