+++ pruebas/01-reading-list/offneti/package.json
